requirements.txt
Flask>=3.1,<4.0
Flask-SQLAlchemy>=3.1,<4.0
python-dotenv>=1.0,<2.0
azure-storage-blob>=12.24,<13.0
azure-ai-vision-imageanalysis>=1.0.0b3,<2.0.0
